# public/swagger.json
{
	"components": {
...
		"schemas": {
			"IPromo": {
				"properties": {
					"promoId": {
						"type": "number",
						"format": "double"
					},
					"promoName": {
						"type": "string"
					}
				},
				"required": [
					"promoId",
					"promoName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IQuestion": {
				"properties": {
					"questionId": {
						"type": "number",
						"format": "double"
					},
					"questionText": {
						"type": "string"
					},
					"questionCode": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"testId": {
						"type": "number",
						"format": "double"
					},
					"cmd": {
						"type": "string"
					},
					"result": {
						"type": "string"
					},
					"useBdd": {
						"type": "boolean"
					}
				},
				"required": [
					"questionId",
					"questionText",
					"questionCode",
					"score",
					"testId",
					"cmd",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IProfile": {
				"properties": {
					"profileId": {
						"type": "number",
						"format": "double"
					},
					"profileName": {
						"type": "string"
					},
					"profileCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"profileId",
					"profileName",
					"profileCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IProfile_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IProfile"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IProfile.Exclude_keyofIProfile.profileId__": {
				"properties": {
					"profileName": {
						"type": "string"
					},
					"profileCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"profileName",
					"profileCode"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IProfile.profileId_": {
				"$ref": "#/components/schemas/Pick_IProfile.Exclude_keyofIProfile.profileId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IProfileCreate": {
				"$ref": "#/components/schemas/Omit_IProfile.profileId_"
			},
			"IUpdateResponse": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IProfileCreate_": {
				"properties": {
					"profileName": {
						"type": "string"
					},
					"profileCode": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IProfileUpdate": {
				"$ref": "#/components/schemas/Partial_IProfileCreate_"
			},
			"IIndexResponse_IPromo_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IPromo"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IPromo.Exclude_keyofIPromo.promoId__": {
				"properties": {
					"promoName": {
						"type": "string"
					}
				},
				"required": [
					"promoName"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IPromo.promoId_": {
				"$ref": "#/components/schemas/Pick_IPromo.Exclude_keyofIPromo.promoId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IPromoCreate": {
				"$ref": "#/components/schemas/Omit_IPromo.promoId_"
			},
			"Partial_IPromoCreate_": {
				"properties": {
					"promoName": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IPromoUpdate": {
				"$ref": "#/components/schemas/Partial_IPromoCreate_"
			},
			"IIndexResponse_IQuestion_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IQuestion"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IQuestion.Exclude_keyofIQuestion.questionId__": {
				"properties": {
					"questionText": {
						"type": "string"
					},
					"questionCode": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"testId": {
						"type": "number",
						"format": "double"
					},
					"cmd": {
						"type": "string"
					},
					"result": {
						"type": "string"
					},
					"useBdd": {
						"type": "boolean"
					}
				},
				"required": [
					"questionText",
					"questionCode",
					"score",
					"testId",
					"cmd",
					"result"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IQuestion.questionId_": {
				"$ref": "#/components/schemas/Pick_IQuestion.Exclude_keyofIQuestion.questionId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IQuestionCreate": {
				"$ref": "#/components/schemas/Omit_IQuestion.questionId_"
			},
			"Partial_IQuestionCreate_": {
				"properties": {
					"questionText": {
						"type": "string"
					},
					"questionCode": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"testId": {
						"type": "number",
						"format": "double"
					},
					"cmd": {
						"type": "string"
					},
					"result": {
						"type": "string"
					},
					"useBdd": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IQuestionUpdate": {
				"$ref": "#/components/schemas/Partial_IQuestionCreate_"
			},
			"IReponse": {
				"properties": {
					"reponseId": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"questionId": {
						"type": "number",
						"format": "double"
					},
					"testPromoUserId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"reponseId",
					"questionId",
					"testPromoUserId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IReponse_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IReponse"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IReponse.Exclude_keyofIReponse.reponseId__": {
				"properties": {
					"questionId": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"testPromoUserId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"questionId",
					"testPromoUserId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IReponse.reponseId_": {
				"$ref": "#/components/schemas/Pick_IReponse.Exclude_keyofIReponse.reponseId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IReponseCreate": {
				"$ref": "#/components/schemas/Omit_IReponse.reponseId_"
			},
			"Partial_IReponseCreate_": {
				"properties": {
					"questionId": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"testPromoUserId": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IReponseUpdate": {
				"$ref": "#/components/schemas/Partial_IReponseCreate_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/profile": {
			"get": {
				"operationId": "GetProfiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IProfile_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page de profiles.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouveau profile",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileCreate"
							}
						}
					}
				}
			}
		},
		"/profile/{profileId}": {
			"get": {
				"operationId": "ReadProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"description": "Récupérer une profile avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un profile avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un profile",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/promo": {
			"get": {
				"operationId": "GetPromos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IPromo_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page de promos.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer une nouvelle promo",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPromoCreate"
							}
						}
					}
				}
			}
		},
		"/promo/{promoId}": {
			"get": {
				"operationId": "ReadPromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPromo"
								}
							}
						}
					}
				},
				"description": "Récupérer une promo avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "promoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour une promo avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "promoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPromoUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer une promo",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "promoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/question": {
			"get": {
				"operationId": "GetQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IQuestion_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page de questions.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer une nouvelle question",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IQuestionCreate"
							}
						}
					}
				}
			}
		},
		"/question/{questionId}": {
			"get": {
				"operationId": "ReadQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IQuestion"
								}
							}
						}
					}
				},
				"description": "Récupérer une question avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "questionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un question avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "questionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IQuestionUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer une question",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "questionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reponse": {
			"get": {
				"operationId": "GetReponses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IReponse_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page des réponses des utilisateurs  aux questions.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateReponse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer une nouvelle réponse d'un utilisateur à une reponse",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReponseCreate"
							}
						}
					}
				}
			}
		},
		"/reponse/{reponseId}": {
			"get": {
				"operationId": "ReadReponse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReponse"
								}
							}
						}
					}
				},
				"description": "Récupérer une réponse d'un utilisateur à une reponse avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "reponseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateReponse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour une réponse d'un utilisateur à une reponse avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "reponseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReponseUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteReponse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer une réponse d'un utilisateur à une reponse",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "reponseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}